+++ frontend/src/stores/google-maps/markerInstanceStore.ts
